+++ types/config_test.go
@@ -21,7 +21,7 @@ func TestLoadConfig(t *testing.T) {
 
 	assert.Equal(config.HealthCheck.Interval, 120)
 	assert.Equal(config.HealthCheck.Timeout, 10)
-	assert.Equal(config.HealthCheck.CacheTTL, 300)
+	assert.Equal(config.HealthCheck.CacheTTL, int64(300))
 	assert.Equal(config.GetHealthCheckStatusTTL(), int64(0))
 
 	assert.Equal(config.Store, "grpc")
